src/requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.5.0  # https://github.com/andymccurdy/redis-py
hiredis==1.0.1  # https://github.com/redis/hiredis-py
python-chess==0.31.2
coveralls==2.1.1
celery[redis]==4.4.7

# Django
# ------------------------------------------------------------------------------
django==3.0.7  # pyup: < 3.1  # https://www.djangoproject.com/
django-extensions==3.1.1
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/jazzband/django-redis
django-annoying==0.10.6
django-paranoid==0.8
django-currentuser==0.5.*
django-filter==2.4.0
django-anymail==7.1.0  # https://github.com/anymail/django-anymail
# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
django-rest-passwordreset==1.1.0
# Django Channels
channels-redis==3.0.0
